Refactor test macros
- Replace PrintableValueTraits with Repr
- Combine utils into macros header
