+++ src/CMakeLists.txt
@@ -54,6 +54,7 @@ list(APPEND SOURCES
   base/ScopedStreamRedirect.cc
   base/TypeDemangler.cc
   base/detail/Copier.cc
+  base/detail/ReprImpl.cc
   comm/Communicator.cc
   comm/Device.cc
   comm/Logger.cc
