+++ test/gtest/detail/ParallelHandler.cc
@@ -7,7 +7,6 @@
 //---------------------------------------------------------------------------//
 #include "ParallelHandler.hh"
 
-#include "Utils.hh"
 #include "base/ColorUtils.hh"
 #include "comm/Communicator.hh"
 #include "comm/Operations.hh"
